Test organization cleanup.
Files checked in: DataRepo/tests/loaders/base/test_table_column.py DataRepo/tests/loaders/base/test_table_loader.py DataRepo/tests/management/commands/test_load_autoupdates.py DataRepo/tests/management/commands/test_studies_exporter.py DataRepo/tests/models/test_hier_cached_model.py DataRepo/tests/models/test_infusate.py DataRepo/tests/models/test_maintained_model.py DataRepo/tests/models/test_peak_data.py DataRepo/tests/utils/test_infusate_parser.py DataRepo/tests/views/upload/test_submission.py DataRepo/tests/management/commands/test_build_caches.py DataRepo/tests/management/commands/test_export_studies.py DataRepo/tests/models/test_peak_data_label.py
