+++ DataRepo/tests/models/test_hier_cached_model.py
@@ -1,6 +1,5 @@
 from django.core.management import call_command
 
-from DataRepo.management.commands.build_caches import cached_function_call
 from DataRepo.models import (
     Animal,
     MaintainedModel,
@@ -596,40 +595,3 @@ class HierCachedModelTests(TracebaseTestCase):
             a.id,
             msg="The root model record returned by get_root_record is directly related",
         )
-
-
-class BuildCachesTests(TracebaseTestCase):
-    fixtures = ["data_types.yaml", "data_formats.yaml", "lc_methods.yaml"]
-
-    @classmethod
-    def setUpTestData(cls):
-        load_minimum_data()
-        super().setUpTestData()
-
-    def test_cached_function_call(self):
-        c = Animal
-        f = "tracers"
-        a = Animal.objects.all().first()
-        la = Animal.objects.all().last()
-        disable_caching_retrievals()
-        # Get the first and last uncached value
-        uv = getattr(a, f)
-        lv = getattr(la, f)
-
-        enable_caching_retrievals()
-        enable_caching_updates()
-        delete_all_caches()
-
-        # Call cached_function_call to populate all cached values for f
-        cached_function_call(c, f)
-
-        # Try to retrieve those cached values
-        v, s = get_cache(a, f)
-        lv, ls = get_cache(la, f)
-
-        # Ensure the value was cached for both the first and last record
-        # Results are querysets, which never equate, but are equatable as lists
-        self.assertEqual(list(v), list(uv))
-        self.assertTrue(s)
-        self.assertEqual(list(lv), list(uv))
-        self.assertTrue(ls)
